// angular-loja/src/app/clientes/models/cliente-put.ts
export class ClientePut {
    idCliente: number;
    nome: string;
    dataNascimento: string;
    telefone1: string;
    telefone2: string;
    email: string;
    observacao: string;
    logradouro: string;
    numero: string;
    complemento: string;
    condominio: string;
    bairro: string;
    municipio: string;
    estado: string;
    cep: string;
}